requirements.txt
python-dotenv
snowflake-connector-python
pandas
kagglehub
kaggle